split components (#428)
* fix: jotai useSetAtom (#419)

* feat: add github icon (#395)

* feat: add github icon

* feat: update icon's size and gap

* feat: update github icon

* feat: gap and icon position

* feat: go back to last page (#397)

* fix no-data display in period rewards (#391)

* feat(MyReward): update no-data display

* feat: update img

* feat: split components

* feat: split components

* fix: update constants
