+++ src/constants/contracts/request.ts
@@ -1,1 +1,3 @@
 export const STALETIME = 1 * 60 * 1000 //1 minute
+
+export const PAGESIZE = 10
